% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar2.R
\name{auto_bracket}
\alias{auto_bracket}
\title{Adds a bracket automatically with a label to annotate a ggplot2 plot}
\usage{
auto_bracket(plot, matching, y, categories, cex = 2.2, text = matching)
}
\arguments{
\item{plot}{a ggplot2 plot.}

\item{matching}{the category value for the bracket}

\item{y}{the y coordinate of the bracket.}

\item{categories}{the vector of categories (one per sample)}

\item{cex}{the size of the text (default 2.2)}

\item{text}{the value of the text (if it isn't the same as matching)}
}
\description{
For example, you can label groups of samples in a large bar chart.
auto_bracket automatically positions the bracket assuming your list
of categories is in the same order as your samples (which they would
if they were part of a metadata sheet, or a sample_data in phyloseq)
}
\examples{
auto_bracket()
}
