% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_biom.R
\name{deseq2_biom}
\alias{deseq2_biom}
\title{Run DESeq2 on a phyloseq object and return tibble of significantly abundant OTUs}
\usage{
deseq2_biom(biom, taxlevel, group, alpha = 0.01, threshold = 1,
  glom = TRUE)
}
\arguments{
\item{biom}{a phyloseq object.}

\item{taxlevel}{string giving taxonomic level (from tax_table) to compare at.}

\item{group}{string giving column in sample data to compare}

\item{alpha}{p-value cutoff}

\item{threshold}{minimum percentage of OTU to report}

\item{glom}{run tax_glom on biom to get to desired level (FALSE if already at right level)}
}
\description{
Run DESeq2 on a phyloseq object and return tibble of significantly abundant OTUs
}
\examples{
deseq2_biom()
}
