% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_biom.R
\name{donut_biom}
\alias{donut_biom}
\title{Make a donut graph of two taxonomic levels}
\usage{
donut_biom(biom, taxlevel1, taxlevel2, variable, condition, threshold = 0)
}
\arguments{
\item{biom}{a phyloseq object.}

\item{taxlevel1}{string giving taxonomic level (from tax_table) to plot inner pie at.}

\item{taxlevel2}{string giving taxonomic level (from tax_table) to plot outer pie at.}

\item{variable}{variable in sample data to restrict samples to.}

\item{condition}{value of variable to restrict samples to}

\item{threshold}{threshold percentage above which will be included in graph}
}
\description{
Make a donut graph of two taxonomic levels
}
\examples{
donut_biom

}
