% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yams2phyloseq.R
\name{mr2phyloseq}
\alias{mr2phyloseq}
\title{Convert metagenomeSeq to phyloseq}
\usage{
mr2phyloseq(mrexp, unrootedtreeqza = NULL)
}
\arguments{
\item{mrexp}{metagenomeSeq experiment object}

\item{unrootedtreeqza}{optional unrootedtree.qza for Unifrac}
}
\description{
Convert metagenomeSeq to phyloseq
}
\examples{
mr2phyloseq()
}
