% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_deseq2.R
\name{my_boxplot}
\alias{my_boxplot}
\title{Make a boxplot of significant OTUs found with my_deseq2}
\usage{
my_boxplot(biom, taxlevel, condition, results, title = NULL, glom = TRUE,
  printSig = TRUE, cex = 2, colors = NULL)
}
\arguments{
\item{biom}{a phyloseq object.}

\item{taxlevel}{string giving taxonomic level (from tax_table) to plot at.}

\item{condition}{string giving column in sample data to compare}

\item{results}{tibble returned from my_deseq2}

\item{glom}{run tax_glom on biom to get to desired level (FALSE if already at right level)}

\item{printSig}{print p-values on plot}

\item{cex}{font size}

\item{colors}{optional vector of colors}
}
\description{
Make a boxplot of significant OTUs found with my_deseq2
}
\examples{
my_boxplot()
}
