% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_bar.R
\name{my_bracket}
\alias{my_bracket}
\title{Adds a bracket with a label to annotate a ggplot2 plot}
\usage{
my_bracket(plot, text, x, xend, y, cex = 2.2)
}
\arguments{
\item{plot}{a ggplot2 plot.}

\item{text}{the text to label the bracket with.}

\item{x}{the x coordinate of the start of the bracket.}

\item{xend}{the x coordinate of the end of the bracket.}

\item{y}{the y coordinate of the bracket.}

\item{cex}{the size of the text (default 2.2)}
}
\description{
For example, you can label groups of samples in a large bar chart.
More typically you will use auto_bracket() which automatically sets
the x and xend based on a category label.
}
\examples{
my_bracket()
}
