% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_bar.R
\name{my_plot_bar}
\alias{my_plot_bar}
\title{More flexible version of phyloseq's plot_bar}
\usage{
my_plot_bar(biom, fill = NULL, title = NULL, colors = NULL,
  maxTaxa = NULL)
}
\arguments{
\item{biom}{a phyloseq object.}

\item{fill}{string giving taxonomic level (from tax_table) to plot at.}

\item{title}{title of plot}

\item{colors}{vector of color names to use}

\item{maxTaxa}{optional limit to number of taxa to display}
}
\description{
More flexible version of phyloseq's plot_bar
}
\examples{
my_plot_bar()
}
