% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_biom.R
\name{pcoa_biom}
\alias{pcoa_biom}
\title{Make a scatter PCA/PCOA plot of phyloseq object}
\usage{
pcoa_biom(biom, method = "PCoA", distance = "unifrac",
  color_category = NULL, shape_category = NULL, weighted = FALSE,
  title = NULL, colors = NULL, shapes = NULL,
  label_category = NULL, order_category = NULL, permanova = FALSE,
  point_size = 1)
}
\arguments{
\item{biom}{phyloseq object}

\item{method}{data reduction method to use (default PCoA; see phyloseq ordinate)}

\item{distance}{distance measure to use (default unifrac)}

\item{color_category}{variable in sample data (if any) to color points by}

\item{shape_category}{variable in sample data (if any) to distinguish points by symbol}

\item{weighted}{use weighted distance (normally only useful for unifrac)}

\item{title}{title of plot}

\item{shapes}{vector of shapes to use if you don't like the defaults}

\item{label_category}{variable in sample data (if any) to label points}

\item{permanova}{-- include PERMANOVA analysis on color_category}

\item{point_size}{size of points on graph}

\item{color}{vector of colors to use if you don't like the defaults}
}
\description{
Make a scatter PCA/PCOA plot of phyloseq object
}
\examples{
pcoa_biom()
}
