% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_biom.R
\name{pcoa_subset_biom}
\alias{pcoa_subset_biom}
\title{Make a scatter PCA/PCOA plot of phyloseq object with complete biom and subset of samples -- useful for conserved axes}
\usage{
pcoa_subset_biom(biom, method = "PCoA", distance = "unifrac", samples,
  color_category = NULL, shape_category = NULL, weighted = FALSE,
  title = NULL, colors = NULL, shapes = NULL, seed = 100,
  xrange = NULL, yrange = NULL)
}
\arguments{
\item{biom}{phyloseq object}

\item{method}{data reduction method to use (default PCoA; see phyloseq ordinate)}

\item{distance}{distance measure to use (default unifrac)}

\item{samples}{samples to include on plot}

\item{color_category}{variable in sample data (if any) to color points by}

\item{shape_category}{variable in sample data (if any) to distinguish points by symbol}

\item{weighted}{use weighted distance (normally only useful for unifrac)}

\item{title}{title of plot}

\item{shapes}{vector of shapes to use if you don't like the defaults}

\item{seed}{random number generator seed (default 100)}

\item{xrange}{optional range of acceptable points to plot in terms of x values}

\item{yrange}{optional range of acceptable points to plot in terms of y values}

\item{color}{vector of colors to use if you don't like the defaults}
}
\description{
Make a scatter PCA/PCOA plot of phyloseq object with complete biom and subset of samples -- useful for conserved axes
}
\examples{
pcoa_subset_biom
}
