% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yams2phyloseq.R
\name{qiime2MR}
\alias{qiime2MR}
\title{Load QIIME2 data into metagenomeSeq}
\usage{
qiime2MR(tableqza, taxonomyqza, metadata, ppm = TRUE)
}
\arguments{
\item{tableqza}{table.qza qiime2 count data}

\item{taxonomyqza}{taxonomy.qza qiime2 taxonomy data}

\item{metadata}{metadata csv/tsv}
}
\description{
Load QIIME2 data into metagenomeSeq
}
\examples{
qiime2MR()
}
