% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yams2phyloseq.R
\name{yams2phyloseq}
\alias{yams2phyloseq}
\title{Create a phyloseq Object from YAMS16 output}
\usage{
yams2phyloseq(counts, taxa, metadata = NULL, tree = NULL,
  splitSemi = FALSE)
}
\arguments{
\item{counts}{unit_table.tsv count data}

\item{taxa}{taxa_16S_cons.tsv taxonomy data}

\item{metadata}{optional metadata csv/tsv}

\item{tree}{optional tree for Unifrac computation.}
}
\description{
Create a phyloseq Object from YAMS16 output
}
\examples{
yams2phyloseq()
}
