% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_biom.R
\name{donut_biom}
\alias{donut_biom}
\title{Make a donut graph of two taxonomic levels}
\usage{
donut_biom(biom, samples, rank1, rank2, inner_colors, outer_colors,
  radius = c(0.7, 1), threshold = 1, main = "")
}
\description{
biom
group  vector identifying the group for each slice
labels vector of labels for individual slices
col    colors for each group
radius radius for inner and outer pie (usually in [0,1])
threshold threshold percent below which don't display
main title of graph
donut_biom
}
