% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_biom.R
\name{donut_biom}
\alias{donut_biom}
\title{Make a donut graph of two taxonomic levels}
\usage{
donut_biom(biom, samples = NULL, rank1, rank2, inner_colors,
  outer_colors, threshold = 1, title = "", verbose = FALSE)
}
\arguments{
\item{biom}{phyloseq object}

\item{inner_colors}{colors four inner circle}

\item{outer_colors}{colors for outer circle}

\item{threshold}{threshold percent below which don't display}

\item{title}{title of graph}

\item{group}{vector identifying the group for each slice}

\item{labels}{vector of labels for individual slices}
}
\description{
Make a donut graph of two taxonomic levels
}
\examples{
donut_biom()
}
